change macos to be the seperate binary
