+++ gloom-rs/src/main.rs
@@ -8,7 +8,8 @@
 #![allow(unused_variables)]
 */
 extern crate nalgebra_glm as glm;
-use std::sync::{Arc, Mutex};
+use std::sync::{Arc, Mutex, RwLock};
+use std::thread;
 use std::{
    // mem,
    // os::raw::c_void,
@@ -87,67 +88,169 @@ fn main() {
       .with_inner_size(glutin::dpi::LogicalSize::new(INITIAL_SCREEN_W, INITIAL_SCREEN_H));
    let cb = glutin::ContextBuilder::new().with_vsync(true);
    let windowed_context = cb.build_windowed(wb, &el).unwrap();
-   // Acquire the OpenGL Context and load the function pointers.
-   let context = unsafe { windowed_context.make_current().unwrap() };
-   gl::load_with(|symbol| context.get_proc_address(symbol) as *const _);
-
    // Uncomment these if you want to use the mouse for controls, but want it to be confined to the screen and/or invisible.
    // windowed_context.window().set_cursor_grab(true).expect("failed to grab cursor");
    // windowed_context.window().set_cursor_visible(false);
 
    // Set up a shared vector for keeping track of currently pressed keys
    let arc_pressed_keys = Arc::new(Mutex::new(Vec::<VirtualKeyCode>::with_capacity(10)));
+   // Make a reference of this vector to send to the render thread
+   let pressed_keys = Arc::clone(&arc_pressed_keys);
 
    // Set up shared tuple for tracking mouse movement between frames
    let arc_mouse_delta = Arc::new(Mutex::new((0f32, 0f32)));
+   // Make a reference of this tuple to send to the render thread
+   let mouse_delta = Arc::clone(&arc_mouse_delta);
 
    // Set up shared tuple for tracking changes to the window size
    let arc_window_size = Arc::new(Mutex::new((INITIAL_SCREEN_W, INITIAL_SCREEN_H, false)));
+   // Make a reference of this tuple to send to the render thread
+   let window_size = Arc::clone(&arc_window_size);
 
-   // let mut window_aspect_ratio = INITIAL_SCREEN_W as f32 / INITIAL_SCREEN_H as f32;
+   // Spawn a separate thread for rendering, so event handling doesn't block rendering
+   let render_thread = thread::spawn(move || {
+      // Acquire the OpenGL Context and load the function pointers.
+      // This has to be done inside of the rendering thread, because
+      // an active OpenGL context cannot safely traverse a thread boundary
+      let context = unsafe {
+         let c = windowed_context.make_current().unwrap();
+         gl::load_with(|symbol| c.get_proc_address(symbol) as *const _);
+         c
+      };
 
-   // Set up openGL
-   unsafe {
-      gl::Enable(gl::DEPTH_TEST);
-      gl::DepthFunc(gl::LESS);
-      gl::Enable(gl::CULL_FACE);
-      gl::Disable(gl::MULTISAMPLE);
-      gl::Enable(gl::BLEND);
-      gl::BlendFunc(gl::SRC_ALPHA, gl::ONE_MINUS_SRC_ALPHA);
+      // let mut window_aspect_ratio = INITIAL_SCREEN_W as f32 / INITIAL_SCREEN_H as f32;
 
-      // Query OpenGL version
-      let version = util::get_gl_string(gl::VERSION);
-      let version_major: u32 = version.chars().next().unwrap().to_digit(10).unwrap();
-      let version_minor: u32 = version.chars().nth(2).unwrap().to_digit(10).unwrap();
-
-      // MacOS uses OpenGL 4.1, which doesn't support the debug output functionality introduced in OpenGL 4.3
-      if (version_major > 4) || (version_major == 4 && version_minor >= 3) {
+      // Set up openGL
+      unsafe {
+         gl::Enable(gl::DEPTH_TEST);
+         gl::DepthFunc(gl::LESS);
+         gl::Enable(gl::CULL_FACE);
+         gl::Disable(gl::MULTISAMPLE);
+         gl::Enable(gl::BLEND);
+         gl::BlendFunc(gl::SRC_ALPHA, gl::ONE_MINUS_SRC_ALPHA);
          gl::Enable(gl::DEBUG_OUTPUT_SYNCHRONOUS);
          gl::DebugMessageCallback(Some(util::debug_callback), ptr::null());
-      } else {
-         // Simple error checking
-         let error = gl::GetError();
-         if error != gl::NO_ERROR {
-            println!("OpenGL Error: {}", error);
-         }
+
+         // Print some diagnostics
+         println!("{}: {}", util::get_gl_string(gl::VENDOR), util::get_gl_string(gl::RENDERER));
+         println!("OpenGL\t: {}", util::get_gl_string(gl::VERSION));
+         println!("GLSL\t: {}", util::get_gl_string(gl::SHADING_LANGUAGE_VERSION));
       }
 
-      // Print some diagnostics
-      println!("{}: {}", util::get_gl_string(gl::VENDOR), util::get_gl_string(gl::RENDERER));
-      println!("OpenGL\t: {}", util::get_gl_string(gl::VERSION));
-      println!("GLSL\t: {}", util::get_gl_string(gl::SHADING_LANGUAGE_VERSION));
-   }
+      // == // Set up your VAO around here
 
-   let first_frame_time = std::time::Instant::now();
-   let mut previous_frame_time = first_frame_time;
+      // let my_vao = unsafe { 1337 };
+
+      // == // Set up your shaders here
+
+      // Basic usage of shader helper:
+      // The example code below creates a 'shader' object.
+      // It which contains the field `.program_id` and the method `.activate()`.
+      // The `.` in the path is relative to `Cargo.toml`.
+      // This snippet is not enough to do the exercise, and will need to be modified (outside
+      // of just using the correct path), but it only needs to be called once
+
+      /*
+      let simple_shader = unsafe {
+          shader::ShaderBuilder::new()
+              .attach_file("./path/to/simple/shader.file")
+              .link()
+      };
+      */
+
+      // Used to demonstrate keyboard handling for exercise 2.
+      let mut _arbitrary_number = 0.0; // feel free to remove
+
+      // The main rendering loop
+      let first_frame_time = std::time::Instant::now();
+      let mut previous_frame_time = first_frame_time;
+      loop {
+         // Compute time passed since the previous frame and since the start of the program
+         let now = std::time::Instant::now();
+         // let elapsed = now.duration_since(first_frame_time).as_secs_f32();
+         let delta_time = now.duration_since(previous_frame_time).as_secs_f32();
+         previous_frame_time = now;
+
+         // Handle resize events
+         if let Ok(mut new_size) = window_size.lock() {
+            if new_size.2 {
+               context.resize(glutin::dpi::PhysicalSize::new(new_size.0, new_size.1));
+               // window_aspect_ratio = new_size.0 as f32 / new_size.1 as f32;
+               new_size.2 = false;
+               println!("Window was resized to {}x{}", new_size.0, new_size.1);
+               unsafe {
+                  gl::Viewport(0, 0, new_size.0 as i32, new_size.1 as i32);
+               }
+            }
+         }
+
+         // Handle keyboard input
+         if let Ok(keys) = pressed_keys.lock() {
+            for key in keys.iter() {
+               match key {
+                  // The `VirtualKeyCode` enum is defined here:
+                  //    https://docs.rs/winit/0.25.0/winit/event/enum.VirtualKeyCode.html
+                  VirtualKeyCode::A => {
+                     _arbitrary_number += delta_time;
+                  }
+                  VirtualKeyCode::D => {
+                     _arbitrary_number -= delta_time;
+                  }
+
+                  // default handler:
+                  _ => {}
+               }
+            }
+         }
+         // Handle mouse movement. delta contains the x and y movement of the mouse since last frame in pixels
+         if let Ok(mut delta) = mouse_delta.lock() {
+            // == // Optionally access the accumulated mouse movement between
+            // == // frames here with `delta.0` and `delta.1`
+
+            *delta = (0.0, 0.0); // reset when done
+         }
+
+         // == // Please compute camera transforms here (exercise 2 & 3)
+
+         unsafe {
+            // Clear the color and depth buffers
+            gl::ClearColor(0.035, 0.046, 0.078, 1.0); // night sky
+            gl::Clear(gl::COLOR_BUFFER_BIT | gl::DEPTH_BUFFER_BIT);
+
+            // == // Issue the necessary gl:: commands to draw your scene here
+         }
+
+         // Display the new color buffer on the display
+         context.swap_buffers().unwrap(); // we use "double buffering" to avoid artifacts
+      }
+   });
 
    // == //
    // == // From here on down there are only internals.
    // == //
 
+   // Keep track of the health of the rendering thread
+   let render_thread_healthy = Arc::new(RwLock::new(true));
+   let render_thread_watchdog = Arc::clone(&render_thread_healthy);
+   thread::spawn(move || {
+      if render_thread.join().is_err() {
+         if let Ok(mut health) = render_thread_watchdog.write() {
+            println!("Render thread panicked!");
+            *health = false;
+         }
+      }
+   });
+
    // Start the event loop -- This is where window events are initially handled
    el.run(move |event, _, control_flow| {
       *control_flow = ControlFlow::Wait;
+
+      // Terminate program if render thread panics
+      if let Ok(health) = render_thread_healthy.read() {
+         if !*health {
+            *control_flow = ControlFlow::Exit;
+         }
+      }
 
       match event {
          Event::WindowEvent { event: WindowEvent::Resized(physical_size), .. } => {
@@ -200,88 +303,6 @@ fn main() {
             if let Ok(mut position) = arc_mouse_delta.lock() {
                *position = (position.0 + delta.0 as f32, position.1 + delta.1 as f32);
             }
-         }
-         Event::MainEventsCleared => {
-            // == // Set up your VAO around here
-
-            // let my_vao = unsafe { 1337 };
-
-            // == // Set up your shaders here
-
-            // Basic usage of shader helper:
-            // The example code below creates a 'shader' object.
-            // It which contains the field `.program_id` and the method `.activate()`.
-            // The `.` in the path is relative to `Cargo.toml`.
-            // This snippet is not enough to do the exercise, and will need to be modified (outside
-            // of just using the correct path), but it only needs to be called once
-
-            /*
-            let simple_shader = unsafe {
-                shader::ShaderBuilder::new()
-                    .attach_file("./path/to/simple/shader.file")
-                    .link()
-            };
-            */
-
-            // Used to demonstrate keyboard handling for exercise 2.
-            let mut _arbitrary_number = 0.0; // feel free to remove
-
-            // Compute time passed since the previous frame and since the start of the program
-            let now = std::time::Instant::now();
-            // let elapsed = now.duration_since(first_frame_time).as_secs_f32();
-            let delta_time = now.duration_since(previous_frame_time).as_secs_f32();
-            previous_frame_time = now;
-
-            if let Ok(mut new_size) = arc_window_size.lock() {
-               if new_size.2 {
-                  context.resize(glutin::dpi::PhysicalSize::new(new_size.0, new_size.1));
-                  // window_aspect_ratio = new_size.0 as f32 / new_size.1 as f32;
-                  new_size.2 = false;
-                  println!("Window was resized to {}x{}", new_size.0, new_size.1);
-                  unsafe {
-                     gl::Viewport(0, 0, new_size.0 as i32, new_size.1 as i32);
-                  }
-               }
-            }
-
-            // Handle keyboard input
-            if let Ok(keys) = arc_pressed_keys.lock() {
-               for key in keys.iter() {
-                  match key {
-                     // The `VirtualKeyCode` enum is defined here:
-                     //    https://docs.rs/winit/0.25.0/winit/event/enum.VirtualKeyCode.html
-                     VirtualKeyCode::A => {
-                        _arbitrary_number += delta_time;
-                     }
-                     VirtualKeyCode::D => {
-                        _arbitrary_number -= delta_time;
-                     }
-
-                     // default handler:
-                     _ => {}
-                  }
-               }
-            }
-            // Handle mouse movement. delta contains the x and y movement of the mouse since last frame in pixels
-            if let Ok(mut delta) = arc_mouse_delta.lock() {
-               // == // Optionally access the accumulated mouse movement between
-               // == // frames here with `delta.0` and `delta.1`
-
-               *delta = (0.0, 0.0); // reset when done
-            }
-
-            // == // Please compute camera transforms here (exercise 2 & 3)
-
-            unsafe {
-               // Clear the color and depth buffers
-               gl::ClearColor(0.035, 0.046, 0.078, 1.0); // night sky
-               gl::Clear(gl::COLOR_BUFFER_BIT | gl::DEPTH_BUFFER_BIT);
-
-               // == // Issue the necessary gl:: commands to draw your scene here
-            }
-
-            // Display the new color buffer on the display
-            context.swap_buffers().unwrap(); // we use "double buffering" to avoid artifacts
          }
          _ => {}
       }
